requirements.txt
Flask==3.1.0
pytest
pydantic==2.11.4
structlog==25.3.0
colorama
python-dotenv